Make a benchmark package
