+++ ovc_mini/benchmark/.gitignore
@@ -1,2 +1,3 @@
 output/
 build/
+__pycache__/
