+++ ovc_mini/benchmark/benchmark/__init__.py
@@ -1,0 +1,4 @@
+"""Benchmark init"""
+
+from benchmark.ssh_session import SSHSession
+from benchmark.base import Benchmark, IperfBench, PayloadBench
